Change to server side flask session
This commit changes the session type so that all data stored by the flask session is stored server side rather than client side. This change was made due to the fact that the 'search' cookie has the potential to be over the client side limit of roughly 4096 bytes, whereas server side doesn't have such a limit so is therefore more appropriate.